+++ src/types/io.ts
@@ -49,23 +49,25 @@ export type PaginationResult<T> = {
   hasMore: boolean;
 };
 
-// Configuration
-export type ProcessConfiguration =
-  | {
-      process?: AOProcess;
-    }
-  | {
-      processId?: string;
-    };
+export type ProcessIdConfig = {
+  processId?: string;
+};
 
-export type EpochInput =
-  | {
-      epochIndex: AoEpochIndex;
-    }
-  | {
-      timestamp: Timestamp;
-    }
-  | undefined;
+export type ProcessConfig = {
+  process?: AOProcess;
+};
+
+export type ProcessConfiguration = ProcessConfig | ProcessIdConfig;
+
+export type EpochTimestampInput = {
+  timestamp: Timestamp;
+};
+
+export type EpochIndexInput = {
+  epochIndex: AoEpochIndex;
+};
+
+export type EpochInput = EpochTimestampInput | EpochIndexInput | undefined;
 
 // AO/ARIO Contract
 export type AoBalances = Record<WalletAddress, number>;
@@ -159,7 +161,7 @@ export type AoEpochData = {
   startTimestamp: Timestamp;
   endTimestamp: Timestamp;
   distributionTimestamp: Timestamp;
-  // @deprecated - use `getDistributions` to get distribution data for a given epoch
+  /** @deprecated - use `getDistributions` to get distribution data for a given epoch **/
   distributions: AoEpochDistributionData;
 };
 export type AoTokenSupplyData = {
@@ -527,8 +529,12 @@ export interface AoARIORead {
   getCurrentEpoch(): Promise<AoEpochData>;
   getPrescribedObservers(epoch?: EpochInput): Promise<AoWeightedObserver[]>;
   getPrescribedNames(epoch?: EpochInput): Promise<string[]>;
-  getObservations(epoch?: EpochInput): Promise<AoEpochObservationData>;
-  getDistributions(epoch?: EpochInput): Promise<AoEpochDistributionData>;
+  getObservations(
+    epoch?: EpochInput,
+  ): Promise<AoEpochObservationData | undefined>;
+  getDistributions(
+    epoch?: EpochInput,
+  ): Promise<AoEpochDistributionData | undefined>;
   getTokenCost({
     intent,
     type,
@@ -570,7 +576,19 @@ export interface AoARIORead {
     address: WalletAddress;
   }): Promise<AoRedelegationFeeInfo>;
   getGatewayRegistrySettings(): Promise<AoGatewayRegistrySettings>;
+  getAllDelegates(
+    params?: PaginationParams<AoAllDelegates>,
+  ): Promise<PaginationResult<AoAllDelegates>>;
 }
+
+export type AoAllDelegates = {
+  address: WalletAddress;
+  gatewayAddress: WalletAddress;
+  delegatedStake: number;
+  startTimestamp: Timestamp;
+  vaultedStake: number;
+  cursorId: string;
+};
 
 export interface AoARIOWrite extends AoARIORead {
   // write interactions
@@ -673,13 +691,13 @@ export interface AoARIOWrite extends AoARIORead {
 // Typeguard functions
 export function isProcessConfiguration(
   config: object,
-): config is { process: AOProcess } {
+): config is Required<ProcessConfiguration> & Record<string, never> {
   return 'process' in config;
 }
 
 export function isProcessIdConfiguration(
   config: object,
-): config is { processId: string } {
+): config is Required<ProcessIdConfig> & Record<string, never> {
   return (
     'processId' in config &&
     typeof config.processId === 'string' &&