+++ tests/e2e/web/src/App.tsx
@@ -3,7 +3,9 @@ import {
   ANT_REGISTRY_ID,
   AOProcess,
   ARIO,
+  ContractSigner,
   arioDevnetProcessId,
+  createAoSigner,
 } from '@ar.io/sdk/web';
 import { connect } from '@permaweb/aoconnect';
 import { useEffect, useState } from 'react';
@@ -12,6 +14,15 @@ import remarkGfm from 'remark-gfm';
 
 import './App.css';
 
+// just validating that the default ARIO works in web context
+const defaultArIO = ARIO.init();
+// validating that the writeable ARIO works in web context
+const writeableArIO = ARIO.init({
+  signer: createAoSigner({} as ContractSigner),
+});
+// validating that the ANT registry works in web context
+const antRegistry = ANTRegistry.init();
+// validating that the ARIO works in web context with a process
 const ario = ARIO.init({
   process: new AOProcess({
     processId: process.env.ARIO_PROCESS_ID || arioDevnetProcessId,
@@ -20,7 +31,7 @@ const ario = ARIO.init({
     }),
   }),
 });
-const antRegistry = ANTRegistry.init();
+
 function App() {
   const [contract, setContract] = useState<string>('Loading...');
   const [ants, setAnts] = useState<string>('Loading...');